react query-ify the streams
